Explicitly set default StorageInterface implementation in Bootstrap.
The former implementation felt far too magical, considering it was very
confusing to not add storage to the Bootstrap, there was no default
storage in the Bootstrap, yet default storage was indeed set, but a
level or two below Bootstrap. This change makes it very clear what the
default storage implementation is and where it's getting set.